@@ -7,14 +7,19 @@
 		"test": "jest"
 	},
 	"dependencies": {
-		"crypto-browserify": "^3.11.1",
-		"react": "16.0.0-alpha.12",
-		"react-native": "0.47.1",
-		"react-native-fs": "^2.8.1",
-		"react-native-navigation": "^1.1.193",
+		"crypto-browserify": "^3.12.0",
+		"react": "^16.3.2",
+		"react-native": "^0.51.1",
+		"react-native-checkbox": "^2.0.0",
+		"react-native-dotenv": "^0.1.1",
+		"react-native-fs": "^2.9.12",
+		"react-native-gesture-handler": "^1.0.0-alpha.42",
+		"react-native-navigation": "^1.1.444",
+		"react-native-radio-button": "^2.0.1",
 		"react-native-randombytes": "^3.0.0",
-		"react-native-simple-radio-button": "^2.6.1",
-		"realm": "^1.10.1"
+		"react-native-simple-radio-button": "^2.7.1",
+		"react-native-tab-view": "0.0.74",
+		"realm": "^1.13.0"
 	},
 	"devDependencies": {
 		"babel-jest": "20.0.3",
