@@ -1,8 +1,6 @@
-// var MailGun = require('mailgun-es6');
-var RNFS = require('react-native-fs');
-
-var apiKey = "key-fe27c961e1555cfa2e5226a4b5c9e280";
-var domain = "sandboxd93144ba368b41549c37bac4ef7ae5b0.mailgun.org";
+import RNFS from 'react-native-fs';
+import { AsyncStorage } from 'react-native';
+import { SERVER_URL } from 'react-native-dotenv'
 
 class Exporter {
 
@@ -10,22 +8,35 @@ class Exporter {
 
 	}
 
-	writeAndEmailCSV(data, study) {
+	writeAndEmailCSV(data, study, email) {
 		var path = RNFS.DocumentDirectoryPath + '/' + study + '.txt';
 		RNFS.writeFile(path, data, 'utf8')
 			.then((success) => {
 				console.log('FILE WRITTEN!');
 				console.log(path)
+				this.emailCSV(path, email, study);
 			})
 			.catch((err) => {
 				console.log(err.message);
 			});
-
-		this.emailCSV(path)
 	}
 
-	emailCSV(csv) {
+	emailCSV(path, email, study) {
+		var file = new File(path);
+		var formData = new FormData();
+		formData.append('email', email)
+		formData.append('file', file);
+		formData.append('study', study);
+		
+		var request = new XMLHttpRequest();
+		request.open('GET', SERVER_URL);
+		request.send(formData);
 
+		try {
+			await AsyncStorage.setItem('email', email);
+		} catch (error) {
+			console.log('failed to save email')
+		}
 	}
 
 
