@@ -1,11 +1,13 @@
 import { Navigation } from 'react-native-navigation';
 import Home from './screens/Home';
 import Study from './screens/Study';
-import DataView from './screens/DataView';
-import ParticipantStart from './screens/ParticipantStart';
 import NewStudy from './screens/NewStudy';
+import ParticipantStart from './screens/ParticipantStart';
 import SurveyStart from './screens/newsurvey/SurveyStart';
 import SurveyQuestion from './screens/newsurvey/SurveyQuestion';
+import SurveyFinish from './screens/newsurvey/SurveyFinish';
+import SurveyReview from './screens/newsurvey/SurveyReview';
+import Email from './screens/Email';
 
 // Export screens
 export default () => {
@@ -15,7 +17,9 @@ export default () => {
 	Navigation.registerComponent('mobilesus.ParticipantStart', () => ParticipantStart);
 	Navigation.registerComponent('mobilesus.SurveyStart', () => SurveyStart);
 	Navigation.registerComponent('mobilesus.SurveyQuestion', () => SurveyQuestion);
-	Navigation.registerComponent('mobilesus.DataView', () => DataView);
+	Navigation.registerComponent('mobilesus.SurveyFinish', () => SurveyFinish);
+	Navigation.registerComponent('mobilesus.SurveyReview', () => SurveyReview);
+	Navigation.registerComponent('mobilesus.Email', () => Email);
 
 	Navigation.startSingleScreenApp({
 		screen: {
@@ -25,6 +29,9 @@ export default () => {
 			navigatorButtons: {}
 		},
 		passProps: {},
+		appStyle: {
+			navBarNoBorder: true
+		},
 		animationType: 'slide-down'
 	});
 }
